# Use fewer ramp up repetitions when KeepRunningBatch is used