@matta 2021-04-18
Here is an example of the improvement seen in Abseil benchmarks.  Same result, similar iteration count, but the individual benchmarks using `KeepRunningBatch` finish in about half the time wall clock time.

Running all of the Abseil `stl_set_int32_t` benchmarks drops from 96 to 67 seconds.

### With fix:
```
% time ./bazel-bin/absl/container/btree_benchmark --benchmark_filter=BM_stl_set_int32_t               
2021-04-18T09:16:01-07:00
Running ../../bazel-bin/absl/container/btree_benchmark
Run on (4 X 2700 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x2)
  L1 Instruction 32 KiB (x2)
  L2 Unified 256 KiB (x2)
  L3 Unified 3072 KiB (x1)
Load Average: 0.33, 0.60, 0.53
-------------------------------------------------------------------------------
Benchmark                                     Time             CPU   Iterations
-------------------------------------------------------------------------------
BM_stl_set_int32_t_Insert                   868 ns          867 ns       838864
BM_stl_set_int32_t_InsertSorted             298 ns          298 ns      2411734
BM_stl_set_int32_t_InsertSmall              140 ns          140 ns      4956312
BM_stl_set_int32_t_Lookup                   764 ns          764 ns       954752
BM_stl_set_int32_t_FullLookup               162 ns          162 ns      4317977
BM_stl_set_int32_t_Delete                   942 ns          942 ns       838864
BM_stl_set_int32_t_DeleteRange              107 ns          107 ns      6606054
BM_stl_set_int32_t_QueueAddRem             1997 ns         1997 ns      2097152
BM_stl_set_int32_t_MixedAddRem             2293 ns         2293 ns      4194304
BM_stl_set_int32_t_Fifo                     109 ns          109 ns      6469273
BM_stl_set_int32_t_FwdIter                  101 ns          101 ns      6946343
BM_stl_set_int32_t_InsertRangeRandom 1156528950 ns   1156460818 ns            1
BM_stl_set_int32_t_InsertRangeSorted   70176196 ns     70176992 ns           10
./bazel-bin/absl/container/btree_benchmark   67.22s user 0.61s system 99% cpu 1:07.85 total

```

### Without fix:
```
% time ./bazel-bin/absl/container/btree_benchmark --benchmark_filter=BM_stl_set_int32_t
2021-04-18T09:12:29-07:00
Running ./bazel-bin/absl/container/btree_benchmark
Run on (4 X 2700 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x2)
  L1 Instruction 32 KiB (x2)
  L2 Unified 256 KiB (x2)
  L3 Unified 3072 KiB (x1)
Load Average: 0.75, 0.67, 0.52
-------------------------------------------------------------------------------
Benchmark                                     Time             CPU   Iterations
-------------------------------------------------------------------------------
BM_stl_set_int32_t_Insert                   867 ns          867 ns       734006
BM_stl_set_int32_t_InsertSorted             297 ns          297 ns      2306876
BM_stl_set_int32_t_InsertSmall              139 ns          140 ns      5007200
BM_stl_set_int32_t_Lookup                   771 ns          771 ns       986480
BM_stl_set_int32_t_FullLookup               162 ns          162 ns      3782448
BM_stl_set_int32_t_Delete                   920 ns          920 ns       734006
BM_stl_set_int32_t_DeleteRange              108 ns          108 ns      6291480
BM_stl_set_int32_t_QueueAddRem             1973 ns         1973 ns      2097152
BM_stl_set_int32_t_MixedAddRem             2387 ns         2387 ns      4194304
BM_stl_set_int32_t_Fifo                     110 ns          110 ns      6385349
BM_stl_set_int32_t_FwdIter                  102 ns          102 ns      6878566
BM_stl_set_int32_t_InsertRangeRandom 1171164989 ns   1171161832 ns            1
BM_stl_set_int32_t_InsertRangeSorted   70163345 ns     70160951 ns           10
./bazel-bin/absl/container/btree_benchmark  2>&1  95.75s user 0.56s system 99% cpu 1:36.33 total
```

### With fix (completes in 7.4 seconds)
```
% time ./bazel-bin/absl/container/btree_benchmark --benchmark_filter=BM_stl_set_int32_t_DeleteRange --v=3
Running ./bazel-bin/absl/container/btree_benchmark
Run on (4 X 2700 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x2)
  L1 Instruction 32 KiB (x2)
  L2 Unified 256 KiB (x2)
  L3 Unified 3072 KiB (x1)
Load Average: 0.37, 0.49, 0.45
-- LOG(2): Running BM_stl_set_int32_t_DeleteRange for 1
-- LOG(2): Ran in 0.00927572/0.00927806
-- LOG(3): Next iters: 1048580, 10
-- LOG(2): Running BM_stl_set_int32_t_DeleteRange for 1048580
-- LOG(2): Ran in 0.115609/0.115625
-- LOG(3): Next iters: 6349025, 6.05488
-- LOG(2): Running BM_stl_set_int32_t_DeleteRange for 6349025
-- LOG(2): Ran in 0.68657/0.686669
-------------------------------------------------------------------------
Benchmark                               Time             CPU   Iterations
-------------------------------------------------------------------------
BM_stl_set_int32_t_DeleteRange        107 ns          107 ns      6396338
./bazel-bin/absl/container/btree_benchmark 7.27s user 0.06s system 99% cpu 7.348 total
```
### Without fix (completes in 13.5 seconds)
```
% ./bazel-bin/absl/container/btree_benchmark --benchmark_filter=BM_stl_set_int32_t_DeleteRange --v=3
Running ./bazel-bin/absl/container/btree_benchmark
Run on (4 X 2700 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x2)
  L1 Instruction 32 KiB (x2)
  L2 Unified 256 KiB (x2)
  L3 Unified 3072 KiB (x1)
Load Average: 0.18, 0.42, 0.43
-- LOG(2): Running BM_stl_set_int32_t_DeleteRange for 1
-- LOG(2): Ran in 0.00944516/0.00944853
-- LOG(3): Next iters: 10, 10
-- LOG(2): Running BM_stl_set_int32_t_DeleteRange for 10
-- LOG(2): Ran in 0.00932258/0.00932574
-- LOG(3): Next iters: 100, 10
-- LOG(2): Running BM_stl_set_int32_t_DeleteRange for 100
-- LOG(2): Ran in 0.00948083/0.00948501
-- LOG(3): Next iters: 1000, 10
-- LOG(2): Running BM_stl_set_int32_t_DeleteRange for 1000
-- LOG(2): Ran in 0.00925071/0.00925326
-- LOG(3): Next iters: 10000, 10
-- LOG(2): Running BM_stl_set_int32_t_DeleteRange for 10000
-- LOG(2): Ran in 0.00948569/0.00948977
-- LOG(3): Next iters: 100000, 10
-- LOG(2): Running BM_stl_set_int32_t_DeleteRange for 100000
-- LOG(2): Ran in 0.0095513/0.00955367
-- LOG(3): Next iters: 1000000, 10
-- LOG(2): Running BM_stl_set_int32_t_DeleteRange for 1000000
-- LOG(2): Ran in 0.113674/0.113695
-- LOG(3): Next iters: 6157970, 6.15797
-- LOG(2): Running BM_stl_set_int32_t_DeleteRange for 6157970
-- LOG(2): Ran in 0.660694/0.660775
-------------------------------------------------------------------------
Benchmark                               Time             CPU   Iterations
-------------------------------------------------------------------------
BM_stl_set_int32_t_DeleteRange        107 ns          107 ns      6186622
./bazel-bin/absl/container/btree_benchmark  --v=3  13.46s user 0.06s system 99% cpu **13.530 total**

```